Improved agent abstraction by moving prompt building methods to local agent
