+++ src/airunner/handlers/llm/agent/agents/base.py
@@ -7,8 +7,9 @@ from typing import (
     Type,
 )
 import datetime
-import platform
+from abc import abstractmethod
 
+from llama_cloud import MessageRole
 from llama_index.core.tools import BaseTool, FunctionTool, ToolOutput
 from llama_index.core.chat_engine.types import AgentChatResponse
 from llama_index.core.base.llms.types import ChatMessage
@@ -22,7 +23,6 @@ from airunner.handlers.llm.agent.rag_mixin import RAGMixin
 from airunner.handlers.llm.agent.external_condition_stopping_criteria import ExternalConditionStoppingCriteria
 from airunner.handlers.llm.agent.tools.chat_engine_tool import ChatEngineTool
 from airunner.handlers.llm.agent.tools.rag_engine_tool import RAGEngineTool
-from airunner.handlers.llm.agent.weather_mixin import WeatherMixin
 from airunner.handlers.llm.storage.chat_store.database import DatabaseChatStore
 from airunner.handlers.llm.agent.memory.chat_memory_buffer import ChatMemoryBuffer
 from airunner.handlers.llm.agent.tools.react_agent_tool import ReActAgentTool
@@ -37,8 +37,7 @@ DEFAULT_MAX_FUNCTION_CALLS = 5
 
 
 class BaseAgent(
-    RAGMixin,
-    WeatherMixin
+    RAGMixin
 ):
     def __init__(
         self,
@@ -50,6 +49,7 @@ class BaseAgent(
     ) -> None:
         RAGMixin.__init__(self)
         self.thread = None
+        self._chat_prompt: str = ""
         self._chatbot = None
         self.llm_settings: LLMSettings = llm_settings
         self._current_tab: Optional[Tab] = None
@@ -59,7 +59,6 @@ class BaseAgent(
         self._do_interrupt: bool = False
         self.history: Optional[List[ChatMessage]] = []
         self._llm: Optional[Type[LLM]] = None
-        self._conversation_summaries: Optional[str] = None
         self._conversation: Optional[Conversation] = None
         self._conversation_id: Optional[int] = None
         self._user: Optional[User] = None
@@ -78,7 +77,7 @@ class BaseAgent(
         self._chat_memory: Optional[ChatMemoryBuffer] = None
         self._current_action: LLMActionType = LLMActionType.NONE
         self._memory: Optional[BaseMemory] = None
-        self.load_rag()
+        # self.load_rag()
         self._react_tool_agent: Optional[ReActAgentTool] = None
         self.default_tool_choice: Optional[Union[str, dict]] = default_tool_choice
         self.max_function_calls: int = max_function_calls
@@ -464,65 +463,6 @@ class BaseAgent(
         return datetime.datetime.now().astimezone().tzname()
     
     @property
-    def _date_time_prompt(self) -> str:
-        return (
-            "Current Date / time information:\n"
-            f"- Date: {self.current_date}.\n"
-            f"- Time: {self.current_time}.\n"
-            f"- Timezone: {self.timezone}.\n"
-        )
-    
-    @property
-    def _operating_system_prompt(self) -> str:
-        return (
-            "Operating system information:\n"
-            f"- System: {platform.system()}\n"
-            f"- Release: {platform.release()}\n"
-            f"- Version: {platform.version()}\n"
-            f"- Machine: {platform.machine()}\n"
-            f"- Processor: {platform.processor()}\n"
-        )
-
-    @property
-    def _speakers_prompt(self) -> str:
-        metadata_prompt = None
-        if self.conversation:
-            data = self.conversation.user_data or []
-            if data is not None:
-                metadata_prompt = ""
-                for item in data:
-                    metadata_prompt += f"-- {item}\n"
-            if metadata_prompt:
-                metadata_prompt = f"- Metadata:\n{metadata_prompt}"
-        metadata_prompt = metadata_prompt or ""
-        return (
-            "User information:\n"
-            f"- Username: {self.username}\n"
-            f"- Location: {self.location_display_name}\n"
-            f"{metadata_prompt}"
-            
-            "Chatbot information:\n"
-            f"- Chatbot name: {self.botname}\n"
-            f"- Chatbot mood: {self.bot_mood}\n"
-            f"- Chatbot personality: {self.bot_personality}\n"
-        )
-    
-    @property
-    def personality_prompt(self) -> str:
-        return (
-            f"{self.botname}'s personality: {self.chatbot.bot_personality}\n"
-            if self.chatbot.use_personality 
-            else ""
-        )
-    
-    @property
-    def mood_prompt(self) -> str:
-        return (
-            "You are a living being, one that has feelings, emotions, a personality with needs and desires.\n"
-            f"{self.botname}'s current mood: {self.bot_mood}\n"
-        ) if self.chatbot.use_mood else ""
-    
-    @property
     def _information_scraper_prompt(self) -> str:
         prompt = (
             "You are an information scraper. You will examine a given text and extract relevant information from it.\n"
@@ -566,103 +506,8 @@ class BaseAgent(
         return prompt
     
     @property
-    def _conversation_summary_prompt(self) -> str:
-        return (
-            f"- Conversation summary:\n{self.conversation.summary}\n"
-            if self.conversation and self.conversation.summary else ""
-        )
-        
-    @property
-    def conversation_summaries(self) -> str:
-        summaries = ""
-        conversations = Conversation.objects.order_by(
-            Conversation.id.desc()
-        )[:5]
-        conversations = list(conversations)
-        conversations = sorted(conversations, key=lambda x: x.id, reverse=True)
-        for conversation in conversations:
-            if conversation.summary:
-                summaries += f"- {conversation.summary}\n"
-        if summaries != "":
-            summaries = f"Recent conversation summaries:\n{summaries}"
-        return summaries
-
-    @property
     def _system_prompt(self) -> str:
-        system_instructions = ""
-        guardrails = ""
-        if (
-            self.chatbot.use_system_instructions and
-            self.chatbot.system_instructions and
-            self.chatbot.system_instructions != ""
-        ):
-            system_instructions = f"Always follow these instructions:\n{self.chatbot.system_instructions}\n"
-        if self.chatbot.use_guardrails and self.chatbot.guardrails_prompt and self.chatbot.guardrails_prompt != "":
-            guardrails = f"Always follow these guardrails:\n{self.chatbot.guardrails_prompt}\n"
-        backstory_prompt = ""
-        if self.chatbot.use_backstory and self.chatbot.backstory and self.chatbot.backstory != "":
-            backstory_prompt = (
-                "------\n"
-                f"**Here is {self.botname}'s backstory:**\n"
-                f"{self.chatbot.backstory}\n"
-                "------\n"
-            )
-        section_prompt = ""
-        if self.current_tab.name == "Canvas":
-            section_prompt = "The user is editing an image using the art tool."
-        elif self.current_tab.name == "Browser" and self.webpage_html != "":
-            section_prompt = (
-                "The user is browsing the web. Here's the content of the webpage:\n"
-                f"{self.webpage_html}"
-            )
-        elif self.current_tab.name == "Document":
-            section_prompt = "The user is working on a document."
-        elif self.current_tab.name == "Game":
-            section_prompt = "The user is playing a game."
-        else:
-            section_prompt = ""
-        prompt = (
-            f"Your name is {self.botname}.\n"
-            f"- The user ({self.username}) is having a conversation with the assistant ({self.botname}).\n"
-            "**Here are the rules that you must always remember:**\n"
-            "- You will ONLY generate dialogue, nothing more.\n"
-            "- Only generate responses in pure dialogue form without including any actions, descriptions or stage "
-            "directions in parentheses. Only return spoken words.\n"
-            "- Do not generate redundant dialogue. Examine the conversation and context close and keep responses "
-            "interesting and creative.\n"
-            "- Do not format the response with the character's name or any other text. Only return the dialogue.\n"
-            f"- As {self.botname}, you will avoid repeating previous dialogue.\n"
-            f"- {self.botname} and {self.username} are both humans. They are not AI or bots.\n"
-            f"- As {self.botname}, you will always respond in a way that is consistent with {self.botname}'s emotions, "
-            f"personality, and the context of the conversation.\n"
-            f"- As {self.botname}, you will respond emotionally, persuasively, and realistically.\n"
-            f"- As {self.botname}, you will not ask how you they can provide assistance, or how they can help.\n"
-            f"- As {self.botname}, you must respond and talk like a real human user.\n"
-            f"- As {self.botname}, you can ask questions, make statements, and provide information.\n"
-            f"- As {self.botname}, you should respond in character with emotionally charged responses that are "
-            f"consistent with your personality and mood.\n"
-            f"{backstory_prompt}"
-            f"{system_instructions}"
-            f"{guardrails}"
-            f"{self.conversation_summaries}"
-            "------\n"
-            "**Here is more context that you can use to generate a response:**\n"
-            f"{self.personality_prompt}"
-            f"{self.mood_prompt}"
-            f"{self._date_time_prompt}"
-            f"{self._operating_system_prompt}"
-            f"{self._speakers_prompt}"
-            f"{self._weather_prompt}"
-            f"{self._conversation_summary_prompt}"
-            f"------\n"
-            "**More information about the current conversation:**\n"
-            f"The conversation is between user ({self.username}) and assistant ({self.botname}).\n"
-            f"The conversation started on {self.conversation.timestamp}.\n"
-            f"{section_prompt}"
-        )
-        prompt = prompt.replace("{{ username }}", self.username)
-        prompt = prompt.replace("{{ botname }}", self.botname)
-        return prompt
+        return self.build_system_prompt()
     
     @property
     def _update_user_data_prompt(self) -> str:
@@ -754,6 +599,10 @@ class BaseAgent(
     def chat_memory(self, value: Optional[ChatMemoryBuffer]):
         self._chat_memory = value
 
+    @abstractmethod
+    def build_system_prompt(self) -> str:
+        pass
+
     def on_load_conversation(self, data: Optional[Dict] = None):
         data = data or {}
         conversation_id = data.get("conversation_id", None)
@@ -819,6 +668,7 @@ class BaseAgent(
         rag_system_prompt: Optional[str] = None,
         llm_request: Optional[LLMRequest] = None
     ) -> AgentChatResponse:
+        self._chat_prompt = message
         self._complete_response = ""
         self.do_interrupt = False
         message = f"{self.username}: {message}"
@@ -834,6 +684,7 @@ class BaseAgent(
 
         if self.print_llm_system_prompt:
             self.logger.info(self._system_prompt)
+            self.logger.info(llm_request.to_dict())
 
         self._update_system_prompt(system_prompt, rag_system_prompt)
